- Implement test script for creating training set and training LOPQ model
Add train_lopq to perform training.

Next:

- Implement location net and clustering using PQKMeans

- Implement training detection model using YOLO

- Implement fine-tuning / training face recognition models

+++ tests/clean.py
@@ -19,7 +19,8 @@ if __name__ == '__main__':
     subprocess.check_output(shlex.split('python manage.py flush --no-input'),cwd='../server')
     utils.migrate()
     for dirname in os.listdir(settings.MEDIA_ROOT):
-        shutil.rmtree("{}/{}".format(settings.MEDIA_ROOT,dirname))
+        if 'DS_store' not in dirname:
+            shutil.rmtree("{}/{}".format(settings.MEDIA_ROOT,dirname))
     for log_filename in glob.glob("../logs/*.log"):
         os.remove(log_filename)
     try:
